# [SPARK-17063] [SQL] Improve performance of MSCK REPAIR TABLE with Hive metastore