… sql/core/src/test/scala/org/apache/spark/sql/execution/command/DDLSuite.scala
… sql/hive/src/main/scala/org/apache/spark/sql/hive/client/HiveShim.scala
… sql/hive/src/test/scala/org/apache/spark/sql/hive/execution/HiveDDLSuite.scala