@davies 2016-08-12
@yhuai @sameeragarwal @rxin I had updated the MSCK REPAIR TABLE to list all the leaf files in parallel to avoid the listing in Hive metastore, hopefully this could speed up it a lot (not benchmarked yet).