@rxin 2016-08-13
@davies  Can you create a new JIRA ticket for this change? It is a non-trivial follow-up.