@davies 2016-08-29
Merging into master and 2.0 branch.